make cfx_addr support no_std feature, and optimize code
remove already handled fixme
